키오스크 카트아이템 보내기 진행중

+++ FinalProject/src/main/java/com/ERP/FinalProject/domain/kiosk/entity/SalesRecords.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 public class SalesRecords {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long saleID;
+    private int saleID;
 
     private LocalDateTime saleDate;
