+++ FinalProject/src/main/java/com/ERP/FinalProject/domain/kiosk/repository/SalesRecordsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ERP.FinalProject.domain.kiosk.entity.SalesRecords;
 
-public interface SalesRecordsRepository extends JpaRepository<SalesRecords, Long>{
+public interface SalesRecordsRepository extends JpaRepository<SalesRecords, Integer>{
 
 }
